OP-346: Add integration test for list_known_urefs ffi/API
